% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaSpatialUtils.R
\name{codeToCode}
\alias{codeToCode}
\title{Convert Between ISO2 and ISO3 Country Codes}
\usage{
codeToCode(countryCodes)
}
\arguments{
\item{countryCodes}{vector of country codes to be converted}
}
\value{
A vector of ISO country codes
}
\description{
Converts a vector of ISO 3166-1 alpha-2 codes to the corresponding ISO 3166-1 alpha-3 codes or vice versa.
}
\keyword{conversion}

