% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGACC.R
\name{convertGACC}
\alias{convertGACC}
\title{Convert Geographic Area Coordination Center geojson, as defined by NIFC}
\usage{
convertGACC(nameOnly = FALSE, simplify = TRUE)
}
\arguments{
\item{nameOnly}{logical specifying whether to only return the name without creating the file}

\item{simplify}{logical specifying whether to create "_05" version of the file that is simplified to 5\%}
}
\value{
Name of the dataset being created.
}
\description{
Returns a SpatialPolygonsDataFrame for Geographic Area Coordination Centers (GACCs)
}
\details{
A GACC shapefile is downloaded and converted to a 
SpatialPolygonsDataFrame with additional columns of data. The resulting file will be created
in the spatial data directory which is set with \code{setSpatialDataDir()}.
}
\references{
\url{https://hub.arcgis.com/items/72213d9266eb4aefa4403a1bf21dfd61}
}
\seealso{
setSpatialDataDir
}
\keyword{datagen}
