% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGADM.R
\name{convertGADM}
\alias{convertGADM}
\title{Convert Global Administrative Areas (GADM) Shapefile}
\usage{
convertGADM(
  nameOnly = FALSE,
  simplify = TRUE,
  countryCode = NULL,
  admLevel = 0
)
}
\arguments{
\item{nameOnly}{Logical specifying whether to only return the name without 
creating the file.}

\item{simplify}{Logical specifying whether to create "_05", _02" and "_01" 
versions of the file that are simplified to 5\%, 2\% and 1\%.#'}

\item{countryCode}{ISO-3166-1 alpha-2 country code}

\item{admLevel}{administrative level to be downloaded}
}
\value{
Name of the dataset being created.
}
\description{
Create a SpatialPolygonsDataFrame for Global Administrative Areas.
}
\details{
A Global Administrative Areas shapefile is downloaded and converted to a 
SpatialPolygonsDataFrame with additional columns of data. The resulting file 
will be created in the spatial data directory which is set with 
\code{setSpatialDataDir()}.
}
\note{
From the source documentation:
}
\references{
\url{https://gadm.org/data.html}.
}
\keyword{datagen}
