% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIndianLands.R
\name{convertIndianLands}
\alias{convertIndianLands}
\title{Convert Indian Lands Shapefile}
\usage{
convertIndianLands(nameOnly = FALSE)
}
\arguments{
\item{nameOnly}{logical specifying whether to only return the name without creating the file}
}
\value{
Name of the dataset being created.
}
\description{
A shapefile is downloaded and converted to a SpatialPolygonsDataFrame
 with additional columns of data. The resulting file will be created
in the spatial data directory which is set with \code{setSpatialDataDir()}
}
\details{
The USIndianLands shapefile represents lands administered by the Bureau of Indian Affairs, ie. Indian reservations
 and is compiled by the National Atlas of the United States of America.
}
\references{
\url{https://nationalmap.gov/small_scale/atlasftp.html#indlanp}

\url{https://nationalmap.gov/small_scale/mld/indlanp.html}
}
\seealso{
setSpatialDataDir
}
\keyword{datagen}
