% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertISOCodeTable.R
\name{convertISOCodeTable}
\alias{convertISOCodeTable}
\title{Convert ISO country code Table to Dataframe}
\usage{
convertISOCodeTable()
}
\value{
Dataframe with 247 rows and 4 columns.
}
\description{
Returns a dataframe version of the country code table with the following columns:
\itemize{
  \item{countryName}
  \item{countryCode}
  \item{ISO3}
  \item{ISONumeric}
}
}
\references{
\url{http://www.nationsonline.org/oneworld/country_code_list.htm}
}
\keyword{datagen}
