% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convertLayer.R
\name{convertLayer}
\alias{convertLayer}
\title{Convert Shapefile Layer to Spatial Polygon Dataframe}
\usage{
convertLayer(dsn = "", layerName = "")
}
\arguments{
\item{dsn}{dsn argument to readOGR}

\item{layerName}{layer argument to readOGR}
}
\value{
An object of class \code{SpatialPolygonsDataFrame}
}
\description{
Raw shapefiles are read in using the \code{readOGR()} function from the \pkg{rgdal} package.
Spatial data are reprojected onto a standard projection with \code{"+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs"} before being returned.
}
\keyword{internal}

