% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMTBSBurnArea.R
\name{convertMTBSBurnArea}
\alias{convertMTBSBurnArea}
\title{Convert MTBS Burn Area shapefiles}
\usage{
convertMTBSBurnArea(nameOnly = FALSE, simplify = FALSE)
}
\arguments{
\item{nameOnly}{Logical specifying whether to only return the name without 
creating the file.}

\item{simplify}{Logical specifying whether to create "_05", _02" and "_01" 
versions of the file that are simplified to 5\%, 2\% and 1\%.}
}
\value{
Name of the dataset being created.
}
\description{
Returns a SpatialPolygonsDataFrame for all large wildland fires 
(includes wildfire, wildland fire use, and prescribed fire) in the 
conterminous United States (CONUS), Alaska, Hawaii, and Puerto Rico for the 
period of 1984 through 2017.  All fires reported as greater than 1,000 acres 
in the western U.S. and greater than 500 acres in the eastern U.S. are 
mapped across all ownerships.
}
\note{
Data last updated by USFS: August 22, 2019
}
\references{
\url{https://edcintl.cr.usgs.gov/downloads/sciweb1/shared/MTBS_Fire/data/composite_data/burned_area_extent_shapefile/mtbs_perimeter_data.zip}
}
\keyword{datagen}
