% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPHDs.R
\name{convertPHDs}
\alias{convertPHDs}
\title{Convert Public Health Districts Shapefile}
\usage{
convertPHDs(nameOnly = FALSE)
}
\arguments{
\item{nameOnly}{logical specifying whether to only return the name without creating the file}
}
\value{
Name of the dataset being created.
}
\description{
Returns a SpatialPolygonsDataFrame for Public Health Districts for Washington, 
Oregon, Idaho, and California.
}
\details{
A Public Health Districts shapefile is downloaded and converted to a 
SpatialPolygonsDataFrame with additional columns of data. The resulting file will be created
in the spatial data directory which is set with \code{setSpatialDataDir()}.
}
\references{
\url{http://mazamascience.com/Shapefiles/PHDs.tgz}
}
\seealso{
setSpatialDataDir
}
\keyword{datagen}
