% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTerrestrialEcoregions.R
\name{convertTerrestrialEcoregions}
\alias{convertTerrestrialEcoregions}
\title{Convert Terrestrial Ecoregion Shapefile}
\usage{
convertTerrestrialEcoregions(nameOnly = FALSE, simplify = TRUE)
}
\arguments{
\item{nameOnly}{logical specifying whether to only return the name without creating the file}

\item{simplify}{logical specifying whether to create a "_05" version of the file that is simplified to 5\%}
}
\value{
Name of the dataset being created.
}
\description{
A shapefile is downloaded and converted to a SpatialPolygonsDataFrame
 with additional columns of data. The resulting file will be created
in the spatial data directory which is set with \code{setSpatialDataDir()}
}
\references{
\url{https://www.worldwildlife.org/publications/terrestrial-ecoregions-of-the-world}
}
\keyword{datagen}
