% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTMWorldBorders.R
\name{convertTMWorldBorders}
\alias{convertTMWorldBorders}
\title{Convert world borders shapefile}
\usage{
convertTMWorldBorders(nameOnly = FALSE, simplify = TRUE)
}
\arguments{
\item{nameOnly}{Logical specifying whether to only return the name without
creating the file.}

\item{simplify}{Logical specifying whether to create "_05", _02" and "_01"
versions of the file that are simplified to 5\%, 2\% and 1\%.}
}
\value{
Name of the datasetName being created.
}
\description{
Returns a simple features data frame for world divisions
}
\details{
A world borders shapefile is downloaded and converted to a
simple features data frame with additional columns of data. The resulting file is
created in the spatial data directory which is set with \code{setSpatialDataDir()}.
}
\references{
\url{http://thematicmapping.org/downloads/}
}
\seealso{
setSpatialDataDir

getCountry, getCountryCode
}
