% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTMWorldBordersSimple.R
\name{convertTMWorldBordersSimple}
\alias{convertTMWorldBordersSimple}
\title{Convert (Simple) World Borders Shapefile}
\usage{
convertTMWorldBordersSimple(nameOnly = FALSE, simplify = TRUE)
}
\arguments{
\item{nameOnly}{Logical specifying whether to only return the name without
creating the file.}

\item{simplify}{Logical specifying whether to create "_05", _02" and "_01"
versions of the file that are simplified to 5\%, 2\% and 1\%.}
}
\value{
Name of the dataset being created.
}
\description{
Returns a SpatialPolygonsDataFrame for simple world divisions
}
\details{
A world borders shapefile is downloaded and converted to a
SpatialPolygonsDataFrame with additional columns of data. The resulting file will be created
in the package \code{SpatialDataDir} which is set with \code{setSpatialDataDir()}.

This shapefile is a simplified version of the TMWorldBorders shapefile.
Users may wish to use a higher resolution dataset when plotting.
}
\references{
\url{http://thematicmapping.org/downloads/}
}
\seealso{
setSpatialDataDir

getCountry, getCountryCode
}
\keyword{datagen}
