% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convertUSCensusCounties.R
\name{convertUSCensusCounties}
\alias{convertUSCensusCounties}
\title{Convert US County Borders Shapefile}
\usage{
convertUSCensusCounties(nameOnly = FALSE, ...)
}
\arguments{
\item{nameOnly}{logical specifying whether to only return the name without creating the file}

\item{...}{additional arguments (unused)}
}
\value{
Large SpatialPolygonsDataFrame
}
\description{
A US county borders shapefile is downloaded and converted to a
SpatialPolygonsDataFrame with additional columns of data. The resulting file wil be created
in the package \code{SpatialDataDir} which can be set with \code{setSpatialDataDir()}.
}
\references{
\url{http://www2.census.gov/geo/tiger/GENZ2013}
}
\keyword{shapefile}
\keyword{spatial,}

