% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUSCensusStates.R
\name{convertUSCensusStates}
\alias{convertUSCensusStates}
\title{Convert US Census State Shapefile}
\usage{
convertUSCensusStates(nameOnly = FALSE)
}
\arguments{
\item{nameOnly}{logical specifying whether to only return the name without creating the file}
}
\value{
Name of the dataset being created.
}
\description{
Returns a SpatialPolygonsDataFrame for US States
}
\details{
A US state shapefile is downloaded and converted to a
SpatialPolygonsDataFrame with additional columns of data. The resulting file will be created
in the spatial data directory which is set with \code{setSpatialDataDir()}.
}
\references{
\url{https://www.census.gov/geo/maps-data/data/cbf/cbf_state.html}
}
\seealso{
setSpatialDataDir
}
\keyword{datagen}
