% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convertUSGSHUC.R
\name{convertUSGSHUC}
\alias{convertUSGSHUC}
\title{Convert USGS Hydrologic Unit Shapefile into smaller subsets}
\usage{
convertUSGSHUC(level = "8", nameOnly = FALSE)
}
\arguments{
\item{level}{character string which must be 2, 4, 6, 8, 10 or 14. 12 is not currently supported. .}

\item{nameOnly}{logical specifying whether to only return the name without creating the file}
}
\value{
Name of the dataset being created.
}
\description{
Using the complete WBD dataset downloaded from ftp://rockyftp.cr.usgs.gov/vdelivery/Datasets/Staged/WBD/Shape/
and simplified using mapshaper, this function subsets the complete set of polygons into smaller groupings based
on larger Hydrologic Units. Files are saved into a new folder in the data directory as .RData SpatialPolygonsDataFrame.
The data directory can be set using \code{setSpatialDataDir()}.
}
\note{
This function requires that the data be downloaded from the URL below and
simplified to a much smaller size using mapshaper software and the command
mapshaper filname.shp -simplify 2% -o
#TODO change this below - what should this function return?
}
\references{
\url{ftp://rockyftp.cr.usgs.gov/vdelivery/Datasets/Staged/WBD/Shape}
}
\seealso{
setSpatialDataDir
}
\keyword{datagen}

