% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convertUSGSHUC8.R, R/convertUSGSHUCSubset.R
\name{convertUSGSHUC8}
\alias{convertUSGSHUC8}
\title{Convert USGS Hydrologic Unit Shapefile}
\usage{
convertUSGSHUC8(nameOnly = FALSE)

convertUSGSHUC8(nameOnly = FALSE)
}
\arguments{
\item{nameOnly}{logical specifying whether to only return the name without creating the file}

\item{nameOnly}{logical specifying whether to only return the name without creating the file}
}
\value{
Name of the dataset being created.

Name of the dataset being created.
}
\description{
A hydrologic unit shapefile is downloaded and converted to a
SpatialPolygonsDataFrame. The resulting file will be created in the data directory
which can be set with \code{setSpatialDataDir()}.

Using the complete WBD dataset downloaded from ftp://rockyftp.cr.usgs.gov/vdelivery/Datasets/Staged/WBD/Shape/
and simplified using mapshaper, this function subsets the complete set of polygons into smaller groupings based
on larger Hydrologic Units. Files are saved into a new folder in the data directory as .RData SpatialPolygonsDataFrame.
The data directory can be set using \code{setSpatialDataDir()}.
}
\details{
The HUC250k dataset consists of 8-digit HUCs.
}
\references{
\url{http://water.usgs.gov/GIS/metadata/usgswrd/XML/huc250k.xml}

\url{ftp://rockyftp.cr.usgs.gov/vdelivery/Datasets/Staged/WBD/Shape}
}
\seealso{
setSpatialDataDir

setSpatialDataDir
}
\keyword{datagen}

