% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWorldEEZ.R
\name{convertWorldEEZ}
\alias{convertWorldEEZ}
\title{Convert World Exclusive Economic Zones shapefile}
\usage{
convertWorldEEZ(nameOnly = FALSE)
}
\arguments{
\item{nameOnly}{logical specifying whether to only return the name without 
creating the file}
}
\value{
Name of the dataset being created.
}
\description{
Returns a SpatialPolygonsDataFrame for US counties.
}
\details{
A world EEZ shapefile is converted to a 
SpatialPolygonsDataFrame with additional columns of data. To use this function, 
the file World_EEZ_v11_20191118 must be downloaded into the users spatial 
directory which is set with \code{setSpatialDataDir()}. The resulting file 
will be created in this same spatial data directory.
}
\references{
\url{http://www.marineregions.org/downloads.php}
}
\seealso{
setSpatialDataDir

getCountry, getCountryCode
}
\keyword{datagen}
