% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MazamaSpatialUtils.R
\docType{data}
\name{CountryTable}
\alias{CountryTable}
\title{Dataframe of World Countries}
\format{A Dataframe with 246 rows and 15 columns}
\description{
Data associated with the TM_WORLD_BORDERS_SMPL dataset obtained
from \url{http://thematicmapping.org/downloads/}. Core data include:
\itemize{
\item{countryCode}
\item{countryName}
\item{latitude}
\item{longitude}
}

Additional variables can be seen with \code{names(CountryTable)}
}
\details{
CountryTable.RData is created during initialization and is stored
in the directory specified with \code{setSpatialDataDir()}.
}
\seealso{
setSpatialDataDir
}
\keyword{datasets}

