% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaSpatialUtils.R
\name{countryToCode}
\alias{countryToCode}
\title{Convert country names to country codes}
\usage{
countryToCode(countryNames)
}
\arguments{
\item{countryNames}{Vector of English language country names.}
}
\value{
A vector of ISO 3166-1 alpha-2 codes or NA.
}
\description{
Converts a vector of English country names to the corresponding
ISO 3166-1 alpha-2 codes.
}
