% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaSpatialUtils.R
\name{.removeSpatialDataDir}
\alias{.removeSpatialDataDir}
\title{Remove package data directory}
\usage{
.removeSpatialDataDir()
}
\value{
Silently returns previous value of data directory.
}
\description{
Resets the package data dir to NULL. Used for internal testing.
}
\seealso{
SpatialDataDir

getSpatialDataDir

setSpatialDataDir
}
\keyword{internal}
