% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPolygonID.R
\name{getPolygonID}
\alias{getPolygonID}
\title{Get PolygonID from SPDF of interest}
\usage{
getPolygonID(SPDF)
}
\arguments{
\item{SPDF}{SpatialPolygonsDataFrame of interest}
}
\value{
Vector of PolygonIDs
}
\description{
Extracts the PolygonID column from the SPDF of interest initially generated by \code{organizePolygons()}
}
\keyword{locator}
