% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getSpatialData.R
\name{getSpatialData}
\alias{getSpatialData}
\title{Return Spatial Data Associated with a Set of Locations}
\usage{
getSpatialData(spDF, lon, lat)
}
\arguments{
\item{spDF}{object of class SpatialPolygonsDataFrame}

\item{lon}{vector of longitudes in decimal degrees}

\item{lat}{vector of latitudes in decimal degrees}
}
\value{
Vector or dataframe of data.
}
\description{
Returns a dataframe or vector of data.
}
\details{
Any spatial polygon may be passed in.
}
\keyword{locator}

