% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MazamaSpatialUtils.R
\name{getSpatialDataDir}
\alias{getSpatialDataDir}
\title{Get Package Data Directory}
\usage{
getSpatialDataDir()
}
\value{
absolute path string
}
\description{
Returns the package data directory where spatial data are located.
}
\seealso{
dataDir

setSpatialDataDir
}
\keyword{spatial}

