% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariable.R
\name{getVariable}
\alias{getVariable}
\title{Return SPDF variable at specified locations}
\usage{
getVariable(
  lon,
  lat,
  dataset = NULL,
  variable = NULL,
  countryCodes = NULL,
  allData = FALSE,
  useBuffering = FALSE
)
}
\arguments{
\item{lon}{vector of longitudes in decimal degrees}

\item{lat}{vector of latitudes in decimal degrees}

\item{dataset}{name of spatial dataset to use}

\item{variable}{name of dataframe column to be returned}

\item{countryCodes}{vector of countryCodes}

\item{allData}{logical specifying whether a full dataframe should be returned}

\item{useBuffering}{Logical flag specifying the use of location buffering to 
find the nearest polygon if not target polygon is found.}
}
\value{
Vector or dataframe.
}
\description{
Uses spatial comparison to determine which polygons the 
locations fall into and returns the variable associated with those polygons.
    
If \code{allData = TRUE}, the entire dataframe is returned.
}
\examples{
\dontrun{
library(MazamaSpatialUtils)
setSpatialDataDir("~/Data/Spatial")

loadSpatialData("NaturalEarthAdm1")

lon <- seq(0, 50)
lat <- seq(0, 50)

getVariable(lon, lat, "NaturalEarthAdm1", "gns_lang")
}
}
\seealso{
getSpatialData
}
\keyword{locator}
