% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MazamaSpatialUtils.R
\name{initializeSpatialData}
\alias{initializeSpatialData}
\title{Install Core Datasets}
\usage{
initializeSpatialData()
}
\value{
nothing
}
\description{
Three core datasets must be installed before the functionality
in \pkg{MazamaSpatialUtils} becomes useful. Running \code{initializeSpatialData()} will
install these datasets in the the directory specified by \code{setSpatialDataDir()}
}
\seealso{
setSpatialDataDir
}
\keyword{spatial}

