% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installedSpatialData.R
\name{installedSpatialData}
\alias{installedSpatialData}
\title{List locally installed spatial datasets}
\usage{
installedSpatialData(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical specifying whether or not to print dataset descriptions.}
}
\value{
Vector of dataset names().
}
\description{
Searches the directory set with \code{\link{setSpatialDataDir}}
for locally installed spatial data and returns a list of dataset names that
can be used with \code{\link{loadSpatialData}}.

If \code{verbose = TRUE}, a brief description is provided for each locally
installed dataset.
}
