% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installSpatialData.R
\name{installSpatialData}
\alias{installSpatialData}
\title{Install Spatial Datasets}
\usage{

  installSpatialData(url = "http://mazamascience.com/RData/mazama_spatial_files.tar.gz")
}
\arguments{
\item{url}{location of spatial data .tar.gz file}
}
\value{
Nothing.
}
\description{
Install spatial datasets found  at \code{url} into the directory previously 
set with \code{setSpatialDataDir()}.
}
\seealso{
setSpatialDataDir

loadSpatialData
}
\keyword{environment}

