% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installSpatialData_0.6.R
\name{installSpatialData_0.6}
\alias{installSpatialData_0.6}
\title{Install version 0.6 spatial datasets}
\usage{
installSpatialData_0.6(
  urlBase = "http://data.mazamascience.com/MazamaSpatialUtils/Spatial",
  file = "mazama_spatial_files-0.6.tar.gz"
)
}
\arguments{
\item{urlBase}{location of spatial data files}

\item{file}{name of the tar.gz file containing spatial datasets}
}
\value{
Nothing.
}
\description{
Install spatial datasets found  at \code{url} into the directory
previously set with \code{setSpatialDataDir()}.
}
\keyword{environment}
