% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MazamaSpatialUtils.R
\docType{package}
\name{MazamaSpatialUtils}
\alias{MazamaSpatialUtils}
\alias{MazamaSpatialUtils-package}
\title{Mazama Science spatial data and utility functions.}
\description{
This package contains code to convert various spatial datasets into .RData files
with uniformly named identifiers including:
\itemize{
\item{ countryCode -- ISO 3166-1 alpha-2}
\item{ countryName -- Country name}
\item{ stateCode -- ISO 3166-2 alpha-2}
\item{ timezone -- Olson timezone}
}
Utility functions allow users to determine the country, state, county and timezones
associated with a set of locations, e.g. environmental monitoring sites.

The uniformity of identifiers in the spatial datasets also makes it easy to generate maps
with data from any dataset that uses standard ISO codes for countries or states.
}

