% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizePolygons.R
\name{organizePolygons}
\alias{organizePolygons}
\title{Organize ungrouped polygons}
\usage{
organizePolygons(SPDF, uniqueID, sumColumns = NULL)
}
\arguments{
\item{SPDF}{Object of class SpatialPolygonsDataFrame.}

\item{uniqueID}{Unique identifier to determine which values are duplicated.}

\item{sumColumns}{Vector of column names to be summed.}
}
\value{
SpatialPolygonsDataFrame composed of grouped polygons.
}
\description{
Determines if the SpatialPolygonsDataFrame is grouped. If ungrouped,
function will group duplicated values based on the provided unique identifier.

If \code{sumColumns} is NULL and there are multiple rows that aren't
duplicated but have the same \code{uniqueID}, the original
SpatialPolygonsDataFrame will be returned.
}
\keyword{internal}
