% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run Shiny app example}
\usage{
runExample(appName = "map_app", ...)
}
\arguments{
\item{appName}{app to run}

\item{...}{parameters passed on to \code{runApp()}.}
}
\description{
This function will run the specified shiny app. By default, the 
app will open in a new window. By default, it will run in the foreground in 
your R console, meaning that you have to stop the app to use R again. The 
default app is "map_app" which requires that the WBDHUC datasets and 
\code{NaturalEarthAdm1} be downloaded to 
SpatialDataDir. They can be installed with \code{\link{convertWBDHUC}}.
}
