% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaSpatialUtils-package.R
\name{setSpatialDataDir}
\alias{setSpatialDataDir}
\title{Set package data directory}
\usage{
setSpatialDataDir(dataDir)
}
\arguments{
\item{dataDir}{Directory where spatial datasets are created.}
}
\value{
Silently returns previous value of data directory.
}
\description{
Sets the package data directory where spatial data is located.
If the directory does not exist, it will be created.
}
\seealso{
SpatialDataDir

getSpatialDataDir
}
