% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-convert.R
\name{simplifyAndSave}
\alias{simplifyAndSave}
\title{Save simplified versions of a spatial features dataframe}
\usage{
simplifyAndSave(
  SFDF = NULL,
  datasetName = NULL,
  uniqueIdentifier = NULL,
  dataDir = getSpatialDataDir()
)
}
\arguments{
\item{SFDF}{Simple features dataframe.}

\item{datasetName}{Base name used for \code{SFDF}.}

\item{uniqueIdentifier}{Name of column containing unique polygon identifiers.}

\item{dataDir}{Spatial data directory set with \code{setSpatialDataDir()}.}
}
\value{
Writes data files to disk.
}
\description{
Creates and saves  "_05", _02" and "_01"
versions of \code{SFDF} that are simplified to 5\%, 2\% and 1\%.
.
}
