% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MazamaSpatialUtils.R
\docType{data}
\name{StateTable}
\alias{StateTable}
\title{Dataframe of World States}
\format{A Dataframe with 4647 rows and 52 columns}
\description{
Data associated with the ne_10m_admin_1 dataset obtained
from \url{http//www.naturalearthdata.com/download}. Core data include:
\itemize{
\item{countryCode}
\item{stateCode}
\item{stateName}
}

Additional variables can be seen with \code{names(StateTable)}
}
\details{
StateTable.RData is created during initialization and is stored
in the directory specified with \code{setSpatialDataDir()}.
}
\keyword{datasets}

