% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeByPolygon.R
\name{summarizeByPolygon}
\alias{summarizeByPolygon}
\title{Summarize values by polygon}
\usage{
summarizeByPolygon(longitude, latitude, value, SPDF, polygonName, FUN,
  useBuffering = FALSE)
}
\arguments{
\item{longitude}{vector of longitudes}

\item{latitude}{vector of latitudes}

\item{value}{vector of values at the locations of interest}

\item{SPDF}{SpatialPolygonsDataFrame with polygons used for aggregating}

\item{polygonName}{name of the SPDF column with the polygon names of interest}

\item{FUN}{function to be applied while summarizing (e.g. mean, max, etc.)}

\item{useBuffering}{passed to MazamaSpatialUtils::getSpatialData()}
}
\value{
a dataframe with two columns: polygonName, summary value
}
\description{
Given vectors of longitudes, latitudes and values, this function will summarize given values by
spatial polygon using the \code{FUN} and return a dataframe with polygon names and summary values.
}
\keyword{internal}
