% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MazamaSpatialUtils.R
\docType{data}
\name{TimezoneTable}
\alias{TimezoneTable}
\title{Dataframe of World Timezones}
\format{A Dataframe with 408 rows and 6 columns}
\description{
Data associated with the tz_world dataset obtained
from \url{http://efele.net/maps/tz/world/}. Core data include:
\itemize{
\item{timezone}
\item{countryCode}
\item{latitude}
\item{longitude}
}

Additional variables can be seen with \code{names(TimezoneTable)}
}
\details{
TimezoneTable.RData is created during initialization and is stored
in the directory specified with \code{setSpatialDataDir()}.
}
\keyword{datasets}

