% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaSpatialUtils.R
\name{codeToCountry}
\alias{codeToCountry}
\title{Convert Country Codes to Country Names}
\usage{
codeToCountry(countryCodes)
}
\arguments{
\item{countryCodes}{vector of country codes to be converted}
}
\value{
A vector of English country names or NA.
}
\description{
Converts a vector of ISO 3166-1 alpha-2 codes to the corresponding English names.
}
\keyword{conversion}

