% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convertLayer.R
\name{convertLayer}
\alias{convertLayer}
\title{Convert Shapefile Layer to Spatial Polygon Dataframe}
\usage{
convertLayer(dsn = "", layerName = "")
}
\arguments{
\item{dsn}{dsn argument to readOGR}

\item{layerName}{layer argument to readOGR}
}
\value{
SpatialPolygonsDataFrame
}
\description{
Raw shapfiles are read in using the \code{readOGR()} function from the \pkg{rgdal} package.
Spatial data are reprojected on to a standard projection with \code{"+proj=longlat"} before being returned.
}
\keyword{shapefile}
\keyword{spatial,}

