% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-convert.R
\name{convertLayer}
\alias{convertLayer}
\title{Convert shapefile layer to simple features dataframe}
\usage{
convertLayer(dsn = NULL, layer)
}
\arguments{
\item{dsn}{dsn argument to \code{sf::st_read()}.}

\item{layer}{layer argument to \code{sf::st_read()}.}
}
\value{
An object of class \code{sf}.
}
\description{
Raw shapefiles are read in using \code{\link[sf]{st_read}}.
Spatial data are reprojected onto a standard projection
(\url{https://epsg.io/4269}) before being returned.

If shapefiles have no projection information they are assumed to 'geographic'
coordinates
.
}
