% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMTBSBurnArea.R
\name{convertMTBSBurnArea}
\alias{convertMTBSBurnArea}
\title{Convert MTBS Burn Area Shapefile}
\usage{
convertMTBSBurnArea(nameOnly = FALSE, simplify = TRUE)
}
\arguments{
\item{nameOnly}{Logical specifying whether to only return the name without 
creating the file.}

\item{simplify}{Logical specifying whether to create "_05", _02" and "_01" 
versions of the file that are simplified to 5\%, 2\% and 1\%.}
}
\value{
Name of the dataset being created.
}
\description{
Create a SpatialPolygonsDataFrame for MTBS Burn Areas
}
\details{
A MTBS Burn Area shapefile is downloaded and converted to a 
SpatialPolygonsDataFrame with additional columns of data. The resulting file 
will be created in the spatial data directory which is set with 
\code{setSpatialDataDir()}.

The source data is from 2020.
}
\note{
From the source documentation:

The Monitoring Trends in Burn Severity (MTBS) project assesses the frequency, 
extent, and magnitude (size and severity) of all large wildland fires 
(includes wildfire, wildland fire use, and prescribed fire) in the 
conterminous United States (CONUS), Alaska, Hawaii, and Puerto Rico for the 
period of 1984 through 2018 All fires reported as greater than 1,000 acres in 
the western U.S. and greater than 500 acres in the eastern U.S. are mapped 
across all ownerships. MTBS produces a series of geospatial and tabular data 
for analysis at a range of spatial, temporal, and thematic scales and are 
intended to meet a variety of information needs that require consistent data 
about fire effects through space and time. This map layer is a vector point 
shapefile of the location of all currently inventoried and mappable MTBS 
fires occurring between calendar year 1984 and 2018 for the continental 
United States, Alaska, Hawaii and Puerto Rico. 

The data generated by MTBS will be used to identify national trends in burn 
severity, providing information necessary to monitor the effectiveness of the 
National Fire Plan and Healthy Forests Restoration Act. MTBS is sponsored by 
the Wildland Fire Leadership Council (WFLC), a multi-agency oversight group 
responsible for implementing and coordinating the National Fire Plan and 
Federal Wildland Fire Management Policies. The MTBS project objective is to
provide consistent, 30 meter resolution burn severity data and burned area 
delineations that will serve four primary user groups: 1. National policies 
and policy makers such as the National Fire Plan and WFLC which require 
information about long-term trends in burn severity and recent burn severity 
impacts within vegetation types, fuel models, condition classes, and land 
management activities. 2. Field management units that benefit from mid to 
broad scale GIS-ready maps and data for pre- and post-fire assessment and 
monitoring. Field units that require finer scale burn severity data will also 
benefit from increased efficiency, reduced costs, and data consistency by 
starting with MTBS data. 3. Existing databases from other comparably scaled 
programs, such as Fire Regime and Condition Class (FRCC) within LANDFIRE, 
that will benefit from MTBS data for validation and updating of geospatial 
data sets. 4. Academic and agency research entities interested in fire 
severity data over significant geographic and temporal extents.
}
\references{
\url{https://edcintl.cr.usgs.gov/downloads/sciweb1/shared/MTBS_Fire/data/composite_data/burned_area_extent_shapefile/mtbs_perimeter_data.zip}
}
\seealso{
setSpatialDataDir

getVariable
}
\keyword{datagen}
