% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convertNaturalEarthAdm1.R
\name{convertNaturalEarthAdm1}
\alias{convertNaturalEarthAdm1}
\title{Convert Level 1 (State) Borders Shapefile}
\usage{
convertNaturalEarthAdm1(nameOnly = FALSE, ...)
}
\arguments{
\item{nameOnly}{logical specifying whether to only return the name without creating the file}

\item{...}{additional arguments (unused)}
}
\value{
name of the dataset being created
}
\description{
Returns a SpatialPolygonsDataFrame for a 1st level administrative divisions
}
\details{
Within the \pkg{MazamaSpatialUtils} package the phrase 'state' refers to administrative
divisions beneath the level of the country or nation. This makes sense in the United 'States'. In
other countries this level is known as 'province', 'territory' or some other term.
}
\references{
\url{http//www.naturalearthdata.com/download}
}
\seealso{
setSpatialDataDir

getState, getStateCode
}
\keyword{shapefile}
\keyword{spatial,}

