% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertNaturalEarthAdm0.R,
%   R/convertNaturalEarthAdm1.R
\name{convertNaturalEarthAdm1}
\alias{convertNaturalEarthAdm1}
\title{Convert Level 1 (state) borders shapefile}
\usage{
convertNaturalEarthAdm1()

convertNaturalEarthAdm1()
}
\description{
Returns a simple features data frame for top level administrative divisions.

The full resolution file will be named "NaturalEarthAdm0.rda". In addition,
"_05", _02" and "_01" versions of the file will be created that that are
simplified to 5\%, 2\% and 1\%. Simplified versions will greatly improve the
speed of both searching and plotting.

Returns a simple features data frame for 1st level administrative divisions

The full resolution file will be named "NaturalEarthAdm1.rda". In addition,
"_05", _02" and "_01" versions of the file will be created that that are
simplified to 5\%, 2\% and 1\%. Simplified versions will greatly improve the
speed of both searching and plotting.
}
\details{
A country border shapefile is downloaded and converted to a
simple features data frame with additional columns of data. The resulting file
will be created in the spatial data directory which is set with
\code{setSpatialDataDir()}. The resulting file
will be created in this same spatial data directory.

A state border shapefile is downloaded and converted to a
simple features data frame with additional columns of data. The resulting file
will be created in the spatial data directory which is set with
\code{setSpatialDataDir()}. The resulting file
will be created in this same spatial data directory.

Within the \pkg{MazamaSpatialUtils} package the phrase 'state' refers to
administrative divisions beneath the level of the country or nation. This
makes sense in the United 'States'. In other countries this level is known as
'province', 'territory' or some other term.
}
\references{
https://www.naturalearthdata.com

https://www.naturalearthdata.com
}
