% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertNWSFireZones.R
\name{convertNWSFireZones}
\alias{convertNWSFireZones}
\title{Convert NWS Public Forecast Zones Shapefile}
\usage{
convertNWSFireZones(nameOnly = FALSE, simplify = TRUE)
}
\arguments{
\item{nameOnly}{logical specifying whether to only return the name without creating the file}

\item{simplify}{logical specifying whether to create "_05" version of the file that is simplified to 5\%}
}
\value{
Name of the dataset being created.
}
\description{
Returns a SpatialPolygonsDataFrame for NWS weather forecast zones.
}
\details{
A weather forecast zones shapefile is downloaded and converted to a
SpatialPolygonsDataFrame with additional columns of data. The resulting file will be created
in the spatial data directory which is set with \code{setSpatialDataDir()}.
}
\note{
zoneID is the unique identifier, and is the state code followed by zoneNumber.
}
\references{
\url{https://www.weather.gov/gis/FireZones}
}
\seealso{
setSpatialDataDir
}
\keyword{datagen}
