% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convertTMWorldBorders.R
\name{convertTMWorldBorders}
\alias{convertTMWorldBorders}
\title{Convert World Borders Shapefile}
\usage{
convertTMWorldBorders(nameOnly = FALSE, ...)
}
\arguments{
\item{nameOnly}{logical specifying whether to only return the name without creating the file}

\item{...}{additional arguments (unused)}
}
\value{
name of the dataset being created
}
\description{
A world borders shapefile is downloaded and converted to a
SpatialPolygonsDataFrame with additional columns of data. The resulting file wil be created
in the package \code{SpatialDataDir} which can be set with \code{setSpatialDataDir()}.
}
\references{
\url{http://thematicmapping.org/downloads/}
}
\seealso{
setSpatialDataDir

getCountry, getCountryCode
}
\keyword{shapefile}
\keyword{spatial,}

