% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTMWorldBorders.R
\name{convertTMWorldBorders}
\alias{convertTMWorldBorders}
\title{Convert world borders shapefile}
\usage{
convertTMWorldBorders()
}
\value{
Name of the datasetName being created.
}
\description{
Returns a simple features data frame for world divisions

The full resolution file will be named "TMWorldBorders.rda". In addition,
"_05", _02" and "_01" versions of the file will be created that that are
simplified to 5\%, 2\% and 1\%. Simplified versions will greatly improve the
speed of both searching and plotting.
}
\details{
A world borders shapefile is downloaded and converted to a
simple features data frame with additional columns of data. The resulting file is
created in the spatial data directory which is set with \code{setSpatialDataDir()}.
}
\references{
\url{http://thematicmapping.org/downloads/}
}
