% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTMWorldBorders.R
\name{convertTMWorldBorders}
\alias{convertTMWorldBorders}
\title{Convert World Borders Shapefile}
\usage{
convertTMWorldBorders(nameOnly = FALSE)
}
\arguments{
\item{nameOnly}{logical specifying whether to only return the name without creating the file}
}
\value{
Name of the dataset being created.
}
\description{
Returns a SpatialPolygonsDataFrame for world divisions
}
\details{
A world borders shapefile is downloaded and converted to a 
SpatialPolygonsDataFrame with additional columns of data. The resulting file will be created
in the spatial data directory which is set with \code{setSpatialDataDir()}.
}
\references{
\url{http://thematicmapping.org/downloads/}
}
\seealso{
setSpatialDataDir

getCountry, getCountryCode
}
\keyword{datagen}

