% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convertTMWorldBordersSimple.R
\name{convertTMWorldBordersSimple}
\alias{convertTMWorldBordersSimple}
\title{Convert World Borders Shapefile}
\usage{
convertTMWorldBordersSimple(nameOnly = FALSE, ...)
}
\arguments{
\item{nameOnly}{logical specifying whether to only return the name without creating the file}

\item{...}{additional arguments (unused)}
}
\value{
name of the dataset being created
}
\description{
A world borders shapefile is downloaded and converted to a
SpatialPolygonsDataFrame with additional columns of data. The resulting file wil be created
in the package \code{SpatialDataDir} which can be set with \code{setSpatialDataDir()}.
}
\details{
This shapefile is a greatly simplified version of the TMWorldBorders shapefile and is especially suited
for spacial searches. This is the default dataset used in \code{getCountry()} and \code{getCountryCode()}.
Users may wish to use a higher resolution dataset when plotting.
}
\references{
\url{http://thematicmapping.org/downloads/}
}
\seealso{
setSpatialDataDir

getCountry, getCountryCode
}
\keyword{shapefile}
\keyword{spatial,}

