% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUSFSRangerDistricts.R
\name{convertUSFSRangerDistricts}
\alias{convertUSFSRangerDistricts}
\title{Convert USFS Ranger District shapefiles}
\usage{
convertUSFSRangerDistricts(nameOnly = FALSE, simplify = TRUE)
}
\arguments{
\item{nameOnly}{Logical specifying whether to only return the name without 
creating the file.}

\item{simplify}{Logical specifying whether to create "_05", _02" and "_01" 
versions of the file that are simplified to 5\%, 2\% and 1\%.}
}
\value{
Name of the dataset being created.
}
\description{
Returns a SpatialPolygonsDataFrame for USFS Ranger Districts

The USFSRangerDistricts layer is a polygon shapefile coverage representing 
US Forest Service ranger district administrative boundaries.  Ranger 
districts are sub units of National Forests intended to identify the specific 
organizational units that administer areas.  

The purpose of the data is to provide display, identification, and analysis 
tools for determining current boundary information for Forest Service 
managers, GIS Specialists, and others.

See https://data.fs.usda.gov/geodata/edw/edw_resources/meta/S_USA.RangerDistrict.xml
for more information on this data set.
}
\note{
Refreshed Feb 3, 2020.
}
\references{
\url{'https://data.fs.usda.gov/geodata/edw/edw_resources/shp/S_USA.RangerDistrict.zip'}
}
\keyword{datagen}
