% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convertUSGSHUC8.R
\name{convertUSGSHUC8}
\alias{convertUSGSHUC8}
\title{Convert USGS Hydrologic Unit Shapefile}
\usage{
convertUSGSHUC8(nameOnly = FALSE)
}
\arguments{
\item{nameOnly}{logical specifying whether to only return the name without creating the file}
}
\value{
Name of the dataset being created.
}
\description{
A hydrologic unit shapefile is downloaded and converted to a
SpatialPolygonsDataFrame. The resulting file will be created in the data directory
which can be set with \code{setSpatialDataDir()}.
}
\details{
The HUC250k dataset consists of 8-digit HUCs.
}
\references{
\url{http://water.usgs.gov/GIS/metadata/usgswrd/XML/huc250k.xml}
}
\seealso{
setSpatialDataDir
}
\keyword{datagen}

