% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWorldEEZ.R
\name{convertWorldEEZ}
\alias{convertWorldEEZ}
\title{Convert World Exclusive Economic Zones Boundaries Shapefile}
\usage{
convertWorldEEZ(nameOnly = FALSE)
}
\arguments{
\item{nameOnly}{logical specifying whether to only return the name without creating the file}
}
\value{
Name of the dataset being created.
}
\description{
A world EEZ shapefile is downloaded and converted to a 
SpatialPolygonsDataFrame with additional columns of data. The resulting file will be created
in the spatial data directory which is set with \code{setSpatialDataDir()}.
}
\references{
\url{http://www.marineregions.org/downloads.php}
}
\seealso{
setSpatialDataDir

getCountry, getCountryCode
}
\keyword{datagen}

