% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convertWorldTimezones.R
\name{convertWorldTimezones}
\alias{convertWorldTimezones}
\title{Create Timezone Dataset}
\usage{
convertWorldTimezones(nameOnly = FALSE, ...)
}
\arguments{
\item{nameOnly}{logical specifying whether to only return the name without creating the file}

\item{...}{additional arguments (unused)}
}
\value{
name of the dataset being created
}
\description{
A world timezone shapefile is downloaded from \url{http://efele.net/maps/tz/world/}
and converted to a SpatialPolygonsDataFrame with additional columns of data. The resulting file wil be created
in the package \code{SpatialDataDir} which can be set with \code{setSpatialDataDir()}.
}
\seealso{
setSpatialDataDir

convertWikipediaTimezoneTable
}
\keyword{shapefile}
\keyword{spatial,}

