% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaSpatialUtils.R
\name{countryToCode}
\alias{countryToCode}
\title{Convert Country Names to Country Codes}
\usage{
countryToCode(countryNames)
}
\arguments{
\item{countryNames}{vector of country names to be converted}
}
\value{
A vector of ISO 3166-1 alpha-2 codes or NA.
}
\description{
Converts a vector of English country names to the corresponding ISO 3166-1 alpha-2 codes.
}
\keyword{conversion}

