% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHUC.R
\name{getHUC}
\alias{getHUC}
\title{Return HUCs at specified locations}
\usage{
getHUC(lon, lat, dataset = "WBDHU10_02", HUCs = NULL,
  allData = FALSE)
}
\arguments{
\item{lon}{vector of longitudes in decimal degrees}

\item{lat}{vector of latitudes in decimal degrees}

\item{dataset}{name of spatial dataset to use}

\item{HUCs}{vector of Hydrologic Unit Codes}

\item{allData}{logical specifying whether a full dataframe should be returned}
}
\value{
Vector of HUC identifiers.
}
\description{
Uses spatial comparison to determine which HUC polygons the 
locations fall into and returns the HUC identifier strings for those polygons.
    
If \code{allData = TRUE}, additional data is returned.
}
\seealso{
getSpatialData
}
\keyword{locator}
