% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHUC.R
\name{getHUC}
\alias{getHUC}
\title{Return HUCs at Specified Locations}
\usage{
getHUC(lon, lat, SPDF, HUCs = NULL, allData = FALSE)
}
\arguments{
\item{lon}{vector of longitudes in decimal degrees}

\item{lat}{vector of latitudes in decimal degrees}

\item{SPDF}{spatial polygons dataset of HUCs}

\item{HUCs}{vector of Hydrologic Unit Codes}

\item{allData}{logical specifying whether to return a full dataframe}
}
\value{
Vector of HUC identifiers.
}
\description{
Uses spatial comparison to determine which HUC polygons the 
    locations fall into and returns the HUC identifier strings for those polygons.
    
    If \code{allData=TRUE}, additional data is returned.
}
\seealso{
getSpatialData
}
\keyword{locator}

