% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHUCName.R
\name{getHUCName}
\alias{getHUCName}
\title{Return HUC Names at Specified Locations}
\usage{
getHUCName(lon, lat, dataset = "WBDHU10-ms", HUCs = NULL, allData = FALSE)
}
\arguments{
\item{lon}{vector of longitudes in decimal degrees}

\item{lat}{vector of latitudes in decimal degrees}

\item{dataset}{name of spatial dataset to use}

\item{HUCs}{vector of Hydrologic Unit Codes}

\item{allData}{logical specifying whether to return a full dataframe}
}
\value{
Vector of HUC names.
}
\description{
Uses spatial comparison to determine which HUC polygons the 
    locations fall into and returns the HUC names for those polygons.
    
    If \code{allData=TRUE}, additional data is returned.
}
\seealso{
getSpatialData
}
\keyword{locator}

