% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPolygonID.R
\name{getPolygonID}
\alias{getPolygonID}
\title{Get polygonID from SFDF of interest}
\usage{
getPolygonID(SFDF)
}
\arguments{
\item{SFDF}{Spatial polygons dataset of interest.}
}
\value{
Vector of polygon identifiers.
}
\description{
Extracts the the vector of unique polygon identifiers from
\code{SFDF}.

This function is useful when writing code to aggregate data by polygon and
calculate per-polygon statistics. Each unique simple features data frame will
have a different set of data columns but each is guaranteed to have a column
named \code{polygonID} that uniquely identifies each polygon.

This allows us to write code that aggregates by polygon without having to
know whether the polygons represent, countries, timezones or HUCs, etc.
}
\keyword{locator}
