% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPolygonID.R
\name{getPolygonID}
\alias{getPolygonID}
\title{Get PolygonID from SPDF of Interest}
\usage{
getPolygonID(SPDF)
}
\arguments{
\item{SPDF}{spatial polygons datset of interest}
}
\value{
Vector of polygon identifiers.
}
\description{
Extracts the the vector of unique polygon identifiers from \code{SPDF}.

This function is useful when writing code to aggregate data by polygon and
calculate per-polygon statistics. Each unique SpatialPolygonsDataFrame will have
a different set of data columns but each is guaranteed to have a column named
\code{polygonID} that uniquely identifies each polygon.

This allows us to write code that aggregates by polygon without having to know
whether the polygons represent, countries, timezones or HUCs, etc.
}
\keyword{locator}
