% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaSpatialUtils.R
\name{getSpatialDataDir}
\alias{getSpatialDataDir}
\title{Get Package Data Directory}
\usage{
getSpatialDataDir()
}
\value{
Absolute path string.
}
\description{
Returns the package data directory where spatial data is located.
}
\seealso{
dataDir

setSpatialDataDir
}
\keyword{environment}

