% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installSpatialData.R
\name{installSpatialData}
\alias{installSpatialData}
\title{Install spatial datasets}
\usage{
installSpatialData(
  dataset = NULL,
  urlBase = "http://data.mazamascience.com/MazamaSpatialUtils/Spatial_0.8"
)
}
\arguments{
\item{dataset}{Name of spatial dataset to install.}

\item{urlBase}{Location of spatial data files.}
}
\value{
If \code{pattern = NULL} a vector of dataset names.
}
\description{
Install spatial datasets found  at \code{url} into the directory
previously set with \code{setSpatialDataDir()}.

If \code{pattern = NULL} (default), available datasets will be displalyed..
}
\keyword{environment}
