% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installSpatialData.R
\name{installSpatialData}
\alias{installSpatialData}
\title{Install a Named Spatial Dataset}
\usage{
installSpatialData(dataset = NULL, verbose = FALSE, ...)
}
\arguments{
\item{dataset}{name of dataset}

\item{verbose}{logical flag controlling detailed progress statements}

\item{...}{additional arguments needed by some \code{convert~} functions}
}
\value{
Character name of the installed dataset.
}
\description{
Install a named spatial dataset.
}
\keyword{environment}

