% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaSpatialUtils.R
\name{iso2ToIso3}
\alias{iso2ToIso3}
\title{Convert from ISO2 to ISO3 country codes}
\usage{
iso2ToIso3(countryCodes)
}
\arguments{
\item{countryCodes}{vector of country codes to be converted}
}
\value{
A vector of ISO3 country codes
}
\description{
Converts a vector of ISO 3166-1 alpha-2 codes to the 
corresponding ISO 3166-1 alpha-3 codes.
}
\keyword{conversion}
