% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSpatialData.R
\name{loadSpatialData}
\alias{loadSpatialData}
\title{Load a Named Spatial Dataset into the Global Environment}
\usage{
loadSpatialData(dataset = NULL)
}
\arguments{
\item{dataset}{name of dataset}
}
\description{
Load a named spatial dataset. The package \code{SpatialDataDir} will be
searched for the specified dataset.  Set this location with \code{setSpatialDataDir()}.
}
\seealso{
getSpatialDataDir

setSpatialDataDir
}
\keyword{environment}

