% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizePolygons.R
\name{organizePolygons}
\alias{organizePolygons}
\title{Organize Ungrouped Polygons}
\usage{
organizePolygons(dataset, uniqueID, sumColumns = NULL)
}
\arguments{
\item{dataset}{object of class SpatialPolygonsDataFrame}

\item{uniqueID}{unique identifier to determine which values are duplicated}

\item{sumColumns}{vector of column names to be summed}
}
\value{
SpatialPolygonsDataFrame composed of grouped polygons.
}
\description{
Determines if the SpatialPolygonsDataFrame is grouped. If ungrouped, 
function will group duplicated values based on the provided unique identifier. 

If sumColumns is NULL and there are multiple rows that aren't duplicated but have
the same 'uniqueID', the original SpatialPolygonsDataFrame will be returned.
}
\examples{
SPDF <- organizePolygons(SimpleTimezones, 'timezone', NULL)
}
\keyword{internal}

