% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaSpatialUtils.R
\name{setSpatialDataDir}
\alias{setSpatialDataDir}
\title{Set Package Data Directory}
\usage{
setSpatialDataDir(dataDir)
}
\arguments{
\item{dataDir}{directory where spatial datasets are created}
}
\value{
Silently returns previous value of data directory.
}
\description{
Sets the package data directory where spatial data is located.
If the directory does not exist, it will be created.
}
\seealso{
SpatialDataDir

getSpatialDataDir
}
\keyword{environment}

