% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaSpatialUtils.R
\docType{data}
\name{SimpleCountries}
\alias{SimpleCountries}
\title{World Country Polygons}
\format{A SpatialPolygonsDataFrame with 246 elements and 7 columns of data.}
\description{
SimpleCountries is a simplified world borders dataset suitable for global maps
and quick spatial searches. This dataset is distributed with the package and is used by
default whenever a dataset with country polygons is required.
}
\details{
This dataset is equivalent to TMWorldBordersSimple but with fewer columns of data.
}
\seealso{
convertTMWorldBordersSimple
}
\keyword{datasets}

