% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaSpatialUtils.R
\docType{data}
\name{SimpleCountriesEEZ}
\alias{SimpleCountriesEEZ}
\title{World Country EEZ Polygons}
\format{A SpatialPolygonsDataFrame with 261 elements and 6 columns of data.}
\description{
SimpleCountriesEEZ is a simplified world borders dataset with a  
200 mile coastal buffer corresponding to Exclusive Economic Zones, suitable for 
quick spatial searches. This dataset is distributed with the package and is used by
default whenever a dataset with country polygons is required.
}
\details{
This dataset is equivalent to EEZCountries but with fewer columns of data.
}
\seealso{
convertEEZCountries
}
\keyword{datasets}
