% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaSpatialUtils.R
\docType{data}
\name{SimpleTimezones}
\alias{SimpleTimezones}
\title{World timezone polygons}
\format{A SpatialPolygonsDataFrame with 1106 elements and 6 columns of data.}
\description{
SimpleTimezones is a simplified world timezones dataset suitable for global maps
and quick spatial searches. This dataset is distributed with the package and is used by
default whenever a dataset with timezone polygons is required.
}
\details{
This dataset is a simplified version of WorldTimezones.
}
\seealso{
convertWorldTimezones
}
\keyword{datasets}
