% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaSpatialUtils.R
\docType{data}
\name{US_stateCodes}
\alias{US_stateCodes}
\title{Dataframe of US State Codes}
\format{A dataframe with 51 rows and 6 columns of data.}
\description{
US_stateCodes contains the following columns of data for the
50 United States plus the District of Columbia:
\itemize{
\item{\code{stateCode} -- e.g. MT}
\item{\code{stateName} -- e.g. Montana}
\item{\code{adm1_code} -- e.g. USA-3515}
\item{\code{code_hasc} -- e.g. US.MT}
\item{\code{fips} -- e.g. US30}
}
}
\keyword{datasets}
