% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{US_stateCodes}
\alias{US_stateCodes}
\title{Dataframe of US state codes}
\format{
A dataframe with 52 rows and 3 columns of data.
}
\usage{
US_stateCodes
}
\description{
US_stateCodes the following columns for US states and territories:
\itemize{
\item{\code{stateName} -- English language state name}
\item{\code{stateCode} -- ISO 3166-2 alpha-2}
\item{\code{stateFIPS} -- two-digit FIPS code}
}

This dataset was generated on 2020-06-11 by running:

\preformatted{
}
}
\keyword{datasets}
